FIX error handling, logging, Linux compatibility
Improved error handling and logging in Python scripts.
Fixed all scripts to use LF instead of CRLF for line endings for
compatibility with Linux. This fixed a bug in which an error in
the firs Python script (input_processing.py) would not properly
halt execution of the shell script and subsequent processing.

Changes to be committed:
	modified:   Dockerfile
	modified:   src/scripts/EpizooticRiskModel_Warehouse.R
	modified:   src/scripts/input_processing.py
	modified:   src/scripts/output_processing.py
	modified:   src/scripts/run_scripts.sh